map: Teach it how to resize
When we hit 0.6 (60%) full we'll perform a shallow clone of ourselves
into a new map struct, and attempt to dupe that entire map into the
new map, with the preserved hash keys. This prevents expensive rehashing
as seen with libnica, and instead we're just reindexing.

Additionally we're able to ensure we don't affect the original map
during the resize due to the nature of the resize. If anything failed
we return the original guy, otherwise we'll blit the old map and set
the content to the new map.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>