map: Implement the remove routine
Basically, find out an occupied row for the key/value pair, nuke them
from the map, and leave any of their links intact within the table. If
we insert values later we may well end up reusing these, and individual
frees on chains is hella expensive. We reclaim during resizes and final
teardown of the blob to minimise allocations.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>